backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
google-genai==1.68.*
google-cloud-storage>=2.17
pymupdf==1.25.*
python-dotenv==1.1.*
python-multipart==0.0.*
pydantic==2.10.*
pytest==8.4.*
passlib[bcrypt]==1.7.*
bcrypt>=3.1.0,<5.0.0
itsdangerous==2.2.*
psycopg2-binary>=2.9
pgvector>=0.3
